Lookup function installation
Apparently lookup function installation process is somewhat fragile
right now. It seems that it happens because a connection to an original
db is reused when we iterate over available databases, and the required
pooler appliance is installed into a wrong db (does it happen due to
some golang optimization reshuffling the code?). To prevent that and
make the implementation more robust, put more debug logging about
installation process, close connection before opening a new one and
check to which database we actually connected to. One more minor fix is
to reset stmtBytes, as it was accumulating stuff in between (which does
not lead to any issues, as appliance queries are idempotent, but still).